Restructure modules to avoid package shadowing
