--- modules/monix/src/main/scala/scalacache/monix/modes.scala
@@ -1,14 +1,0 @@
-package scalacache.monix
-
-import monix.eval.Task
-
-import scalacache.cats.effect.CatsEffectInstances
-import scalacache.{Async, Mode}
-
-object modes {
-
-  implicit val task: Mode[Task] = new Mode[Task] {
-    val M: Async[Task] = CatsEffectInstances.asyncForCatsEffectAsync[Task]
-  }
-
-}
